Use proxy object for object lists
Uses a proxy object to pass the class and enum list to the renderers.
This allows looking up the objects by ID, which is more flexible for the
renderers. This also means that the object parent and property ids can
be the actual IDs which a renderer can use to look up any property of
the object instead of the class name.