Rework context implementation
Removes the premapping from the context class and instead allows it to
expand or compact any identifier. Bindings that want to be able to
expand or compact should dump the entire context and implement logic to
do expansion or compaction. The python bindings are re-written to allow
this and are implemented by pulling in the context.py into the generated
code.