tests: Add SHACL model validation
Validates the roundtrip file against the SHACL model (using pyshacl).
Fixed up many things that were wrong with the implementation that this
uncovered.

This test gets a little tricky since pyshacl will actually resolve the
@context in the source file and as such needs to be a really resolvable
URL. This means we have to start up a HTTP web server to host the model
context files and make sure the test files use the ephemeral URL for the
context. The server startup and file generation are efficiently scoped
in pytest so that they only need to happen once.