Revert encoder/decoder changes
Partially reverts the changes made to pass the `Property` object to the
`Encoder` and `Decoder` made in 6fb5601 ("Add xsd:dateTimeStamp
support"). This approach was wrong and inflexible. Instead, pass the
string version of the datetime to the `Encoder` and have the `Decoder`
pass the string for parsing by the `DateTimeProp`