Rework idPropertyName and referencable predicates
Reworks these predicates to be namespaces to a shacl2code specific
namespace (https://jpewdev.github.io/shacl2code/schema#) instead of an
SPDX one. Also rework the refable enumeration to use fully-qualified
IRIs for the values instead of strings.

Finally, the refable and idPropertyNames are inherited from parent
classes if not specified in a given class