python: Order data more carefully
Python 3.7 and later will write out data in dictionaries in the order
the items are added. Be more careful with the order that things are
added to the dictionaries when serializing to make the output more
plesant; specifically "@context" should always be the first key in the
root object, and objects should have "@type" and "@id" as the first two
fields respectively. Also update the inline JSON encoder to match this
ordering