common: BasicJinjaRender: Move more out of render()
Dealing with the model should not be done inside of the render()
function, since different implementations may want to do different
things with it (e.g. output each class in a separate file).

To accommodate this, move the code that deals with the model setup into
output() and render() just focuses on rendering a single template file
to an output.