jsonschema: Add support for shortened named individuals
If a named individual is shortened by the context, it may not be a valid
IRI. Add explicit matches for all of the context-renamed named
individuals that could be used in place of an object.

Note that only context manipulated named individuals need to be checked,
since full IRI name individuals are the same as any other IRI reference.