Fix context handling
The way that contexts were dealt with was incorrect in a lot of ways,
particularly when dealing with global named individuals.

Correct all of these problems and add better test cases for context
expansion and compaction. The new test suite uses `jsonld-cli` from
`npm` to validate the expansion and compaction, which is the same code
used by the official JSON-LD documentation in the JSON-LD playground.

This makes it so that named individuals can be correctly handled, as
long as the property that needs to reference them is `"@type": "@vocab"`
in the context.