golang: Fix type safety assignment
The go bindings had a peculiarity in which an object could be
incorrectly assigned to a property of the wrong type. The reason is that
that interfaces aren't really "types", so any object that implements the
correct methods matches an interface. Practically speaking, this means
that objects may match multiple interfaces and be cross assigned. This
is particularly problematic for sibling objects with no properties,
since they can be cross assigned since there are no distinct properties
to differentiate them.

To solve this, create a private type holder method for each object
interface. This method does nothing, but it does ensure that each
interface has a completely unique signature and cannot be cross assigned
with another.