tests: Remove static expect data
The expected output for tests was originally useful for validating that
output was correct, but since the tests are now quite extensive, they
are not as critical. In addition, having the expected output is
confusing for people trying to submit changes to the project, since it's
easy to accidentally change the wrong files, or forget to run the script
to update them.