Use Collection from typing instead of from collections.abc
1) Use Collection from typing instead of from collections.abc

To keep compatibility with Python 3.8

2) Add comment for "# ignore[arg-type]" at f.write(chunk) in JSONLDSerializer.write() method

.write() expects either str or bytes, depends on how file is open.
But the type checker wants it to be explicit here which one we like to have.

Since chunk is bytes and we open it in "wb" (binary) mode, this is already correct.
So we tell the type checker to ignore this line.

Add this notes to the comment of JSONLDSerializer.write()

Signed-off-by: Arthit Suriyawongkul <arthit@gmail.com>