jsonschema: Use "if" to filter classes
Use "if" to check the type property of a class before attempting to
check it in more detail. This allows classes to be quickly discarded by
the schema validators. It appears that some validators struggle with
bailing out of "allOf" early (in particular, the python jsonschema
validator), so explicitly adding this check yields large parsing
benefits. Do the same for the top level decision between multiple
objects (with an "@graph" property) or a single object.

In testing, this change yielded a 90% reduction in validation
time with the python jsonschema implementation