tests: Mark pyrefly tests as failing
Newer versions of pyrefly appear to be overly pedantic and ignore the
type annotations of functions, preferring to inspect the code for the
actual return types. This breaks rdflib since it frequently has
different type annotations from what functions might actually return
(for example, `rdflib.terms.URIRef`). Since we can't think of a good
work around for this, mark the tests as expected to fail for now.