model: Allow properties to be split
In RDF, the value of an object property is specified using the property
path as the triple predicate. This means it's not possible to have
multiple properties with the same name (path) in an object, since this
would be interpreted as multiple values (e.g. a list) for the same
property.

This means that it is valid for a NodeShape to have multiple property
definitions for the same path in sh:property. In practice, this means
that the object property value must pass validation for all the
specified shape properties.

shacl2code didn't handle this previously and assumed that sh:property
only contained a single validation for each path and would create code
that was incorrect by attempting to create duplicate properties. Fix
this so that it correctly handles duplicate paths in the property list
and merges them together as best as it can.